Add experimental `columnar` module to differential-dataflow
Moves the columnar arrangement / container infrastructure from the
`examples/columnar/columnar_support/` tree into `src/columnar/` as a
public, experimental module. API and internals are explicitly marked as
unstable in the module-level docs; rough edges (`unimplemented!`
`ContainerBytes`, eager-consolidate `leave_dynamic`, single-`U`
`join_function`) are listed up front.

Also generalizes the dynamic-scope helpers: `DynTime<TOuter, T>` is now
parametric (was hardcoded to `Product<u64, PointStamp<u64>>`), and
`leave_dynamic` carries matching bounds plus a `level > 0` assert.

Updates the in-tree consumers: `examples/columnar/main.rs` and
`interactive/examples/ddir_col.rs` switch from path-mounted modules to
`use differential_dataflow::columnar`. `ddir_vec.rs` + `interactive/src/ir.rs`
gain a `RowLike` impl for `SmallVec<A>`, used by the vec-backed ddir
example for its row representation.
