+++ interactive/Cargo.toml
@@ -14,6 +14,7 @@ workspace = true
 [dependencies]
 columnar = { workspace = true }
 differential-dataflow = { workspace = true }
+mimalloc = "0.1.48"
 smallvec = "1.15.1"
 timely = { workspace = true }
