+++ interactive/examples/ddir_col.rs
@@ -1,5 +1,10 @@
 //! DD IR columnar backend: parse, lower, render, execute.
 
+use mimalloc::MiMalloc;
+
+#[global_allocator]
+static GLOBAL: MiMalloc = MiMalloc;
+
 mod types {
     /// A row type backed by Vec but using Strides for columnar bounds.
     /// This ensures uniform-length rows (common in the IR) get compact
@@ -65,9 +70,7 @@ use interactive::parse;
 use interactive::lower;
 use interactive::ir::Program;
 
-#[path = "../../differential-dataflow/examples/columnar/columnar_support.rs"]
-mod columnar_support;
-use columnar_support::*;
+use differential_dataflow::columnar as columnar_support;
 
 mod columnar {
     use super::types::*;
@@ -205,13 +208,10 @@ mod render {
                     use super::columnar::ValColBuilder;
                     let stream = join_traces::<_, _, _, ValColBuilder<DdirUpdate>>(l, r, move |k, v1, v2, t, d1, d2, c| {
                         use differential_dataflow::difference::Multiply;
-                        let k: Row = Columnar::into_owned(k);
-                        let v1: Row = Columnar::into_owned(v1);
-                        let v2: Row = Columnar::into_owned(v2);
                         let d = d1.clone().multiply(d2);
                         let i = [k.as_slice(), v1.as_slice(), v2.as_slice()];
                         let (k2, v2): (Row, Row) = (eval_fields(&proj.key, &i), eval_fields(&proj.val, &i));
-                        c.give((k2, v2, t.clone(), d));
+                        c.give((k2, v2, t, d));
                     });
                     nodes.insert(id, Rendered::Collection(stream.as_collection()));
                 },
@@ -219,21 +219,22 @@ mod render {
                     let Rendered::Arrangement(a) = &nodes[input] else { panic!("Reduce: input must be an Arrangement") };
                     let a = a.clone();
                     let reducer = reducer.clone();
-                    let f: Arc<dyn Fn(&Row, &[(&Row, Diff)], &mut Vec<(Row, Diff)>) + Send + Sync> = match reducer {
-                        interactive::parse::Reducer::Min => Arc::new(|_key, vals, output| { if let Some(min) = vals.iter().map(|(v, _)| *v).min() { output.push((min.clone(), 1)); } }),
+                    type ReduceFn = dyn for<'a> Fn(columnar::Ref<'a, Row>, &[(columnar::Ref<'a, Row>, Diff)], &mut Vec<(Row, Diff)>) + Send + Sync;
+                    let f: Arc<ReduceFn> = match reducer {
+                        interactive::parse::Reducer::Min => Arc::new(|_key, vals, output| {
+                            if let Some(min) = vals.iter().map(|(v, _)| v.as_slice()).min() {
+                                output.push((Row(min.to_vec()), 1));
+                            }
+                        }),
                         interactive::parse::Reducer::Distinct => Arc::new(|_key, _vals, output| { output.push((Row::new(), 1)); }),
-                        interactive::parse::Reducer::Count => Arc::new(|_key, vals, output| { let count: Diff = vals.iter().map(|(_, d)| *d).sum(); if count > 0 { let mut r = Row::new(); r.push(count); output.push((r, 1)); } }),
+                        interactive::parse::Reducer::Count => Arc::new(|_key, vals, output| {
+                            let count: Diff = vals.iter().map(|(_, d)| *d).sum();
+                            if count != 0 { let mut r = Row::new(); r.push(count); output.push((r, 1)); }
+                        }),
                     };
                     let reduced = a.reduce_abelian::<_, ColValBuilder<_,_,_,_>, ColValSpine<_,_,_,_>, _>(
                         "Reduce",
-                        move |k, vals, output| {
-                            let k: Row = Columnar::into_owned(k);
-                            let owned_vals: Vec<(Row, Diff)> = vals.iter().map(|(v, d)| {
-                                (Columnar::into_owned(*v), *d)
-                            }).collect();
-                            let refs: Vec<(&Row, Diff)> = owned_vals.iter().map(|(v, d)| (v, *d)).collect();
-                            f(&k, &refs, output);
-                        },
+                        move |k, vals, output| { f(k, vals, output); },
                         |col, key, upds| {
                             use columnar::{Clear, Push};
                             col.keys.clear();
@@ -327,6 +328,8 @@ fn run(name: &str, stmts: Vec<parse::Stmt>, n_inputs: usize, nodes: u64, edges: 
 
         let mut builders: Vec<OuterBuilder> = (0..n_inputs).map(|_| OuterBuilder::default()).collect();
 
+        let timer = std::time::Instant::now();
+        let timer_load = std::time::Instant::now();
         for e in 0..edges {
             if (e as usize) % peers == index {
                 let input_idx = (e as usize) % inputs.len();
@@ -342,13 +345,13 @@ fn run(name: &str, stmts: Vec<parse::Stmt>, n_inputs: usize, nodes: u64, edges: 
             h.flush();
         }
         while probe.less_than(&1u64) { worker.step(); }
-        let elapsed = std::time::Instant::now();
-        println!("worker {}: {} loaded ({} edges)", index, name, edges);
+        println!("worker {}: {} loaded ({} edges, total {:.2?}, load {:.2?})", index, name, edges, timer.elapsed(), timer_load.elapsed());
 
         let mut cursor = 0u64;
         let mut round = 0u64;
         let limit = rounds.unwrap_or(u64::MAX);
         while round < limit {
+            let timer_round = std::time::Instant::now();
             let time = (round + 2) as u64;
             for _ in 0..batch {
                 let remove_idx = cursor;
@@ -375,10 +378,10 @@ fn run(name: &str, stmts: Vec<parse::Stmt>, n_inputs: usize, nodes: u64, edges: 
 
             round += 1;
             if round % 100 == 0 {
-                println!("worker {}: {} round {} ({:.2?})", index, name, round, elapsed.elapsed());
+                println!("worker {}: {} round {} (total {:.2?}, round {:.2?})", index, name, round, timer.elapsed(), timer_round.elapsed());
             }
         }
-        println!("worker {}: {} done ({} rounds, batch {}, {:.2?})", index, name, round, batch, elapsed.elapsed());
+        println!("worker {}: {} done ({} rounds, batch {}, total {:.2?})", index, name, round, batch, timer.elapsed());
     }).unwrap();
 }
